refactor: app icons 패키지 이동
