+++ app/src/main/java/com/surivalcoding/composerecipeapp/ui/component/RecipeCard.kt
@@ -29,9 +29,9 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.surivalcoding.composerecipeapp.R
 import com.surivalcoding.composerecipeapp.data.mock.fakeSavedRecipe
-import com.surivalcoding.composerecipeapp.ui.AppIcons
 import com.surivalcoding.composerecipeapp.ui.custom.FoodImageBackground
 import com.surivalcoding.composerecipeapp.ui.custom.HoIconToggleButton
+import com.surivalcoding.composerecipeapp.ui.icon.AppIcons
 import com.surivalcoding.composerecipeapp.ui.theme.AppColors
 import com.surivalcoding.composerecipeapp.ui.theme.AppTextStyles
 import com.surivalcoding.composerecipeapp.ui.theme.ComposeRecipeAppTheme
